requirements.txt
certifi==2020.6.20
cffi==1.14.0
click==7.1.2
cryptography==2.9.2
decorator==4.4.2
pycparser==2.20
python-telegram-bot==12.8
six==1.15.0
tornado==6.0.4
